Reordenamiento clases usuarios. Creados metodos para inmuebles

+++ hk_back/hk_back/users/views.py
@@ -10,6 +10,10 @@ from .serializers import UserSerializer
 
 import jwt, datetime
 
+from django.conf import settings
+
+JWT_SECRET = settings.SECRET_KEY
+
 # Create your views here.
 
 # RegisterView class is used to register a new user
@@ -46,7 +50,7 @@ class LoginView(APIView):
             'iat': datetime.datetime.now(datetime.UTC)
         }
 
-        token = jwt.encode(payload, 'secret', algorithm='HS256')
+        token = jwt.encode(payload, JWT_SECRET, algorithm='HS256')
         response = Response()
 
         response.set_cookie(key='jwt', value=token, httponly=True)
@@ -57,7 +61,7 @@ class LoginView(APIView):
         }
         return response
     
-# UserView class is used to get the user details
+# AuthView class is used to get the user details
 class AuthView(APIView):
     def get(self, request):
         token = request.COOKIES.get('jwt')
@@ -65,7 +69,7 @@ class AuthView(APIView):
             raise AuthenticationFailed('Unauthenticated')
 
         try:
-            payload = jwt.decode(token, 'secret', algorithms=['HS256'])
+            payload = jwt.decode(token, JWT_SECRET, algorithms=['HS256'])
         except jwt.ExpiredSignatureError:
             raise AuthenticationFailed('Unauthenticated')
 
@@ -85,14 +89,9 @@ class LogoutView(APIView):
         }
         return response
     
-class UsersView(APIView):
+# UsersListView class is used to get all users
+class UsersListView(APIView):
     parser_classes = [MultiPartParser, FormParser, JSONParser]
-
-    def get_object(self, pk):
-        try:
-            return User.objects.get(pk=pk)
-        except User.DoesNotExist:
-            raise None
 
     def get(self, request):
         token = request.COOKIES.get('jwt')
@@ -105,6 +104,10 @@ class UsersView(APIView):
             status=status.HTTP_200_OK,
             data=serializer.data
         )
+    
+# UserCreateView class is used to create a new user
+class UserCreateView(APIView):
+    parser_classes = [MultiPartParser, FormParser, JSONParser]
     
     def post(self, request):
         serializer = UserSerializer(data=request.data)
@@ -141,14 +144,19 @@ class UsersView(APIView):
         }
         return response
     
+# UserDetailView class is used to retrieve, update or delete a specific user
+class UserDetailView(APIView):
+    parser_classes = [MultiPartParser, FormParser, JSONParser]
+
+    def get_object(self, pk):
+        try:
+            return User.objects.get(pk=pk)
+        except User.DoesNotExist:
+            raise None
+    
     def patch(self, request, id):
         user = self.get_object(id)
 
-        token = request.COOKIES.get('jwt')
-
-        if not token:
-            raise AuthenticationFailed('Unauthenticated')
-        
         if user is None:
             return Response(
                 status=status.HTTP_404_NOT_FOUND,
@@ -156,6 +164,12 @@ class UsersView(APIView):
                     'message': 'User not found'
                 }
             )
+        
+        token = request.COOKIES.get('jwt')
+
+        if not token:
+            raise AuthenticationFailed('Unauthenticated')
+        
         
         serializer = UserSerializer(user, data=request.data, partial=True)
 
@@ -180,11 +194,6 @@ class UsersView(APIView):
     def delete(self, request, id):
         user = self.get_object(id)
 
-        token = request.COOKIES.get('jwt')
-
-        if not token:
-            raise AuthenticationFailed('Unauthenticated')
-        
         if user is None:
             return Response(
                 status=status.HTTP_404_NOT_FOUND,
@@ -192,6 +201,11 @@ class UsersView(APIView):
                     'message': 'User not found'
                 }
             )
+        
+        token = request.COOKIES.get('jwt')
+
+        if not token:
+            raise AuthenticationFailed('Unauthenticated')
         
         user.delete()
         response = Response()
